Remove client side caching and move url validation to server side
